+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.8.0
+
+- Support folder level validation.
+- Add rule INVALID_FILE_LOCATION to validate if management plane swagger in 'resource-manager' folder
+
 ## 0.7.2
 
 - Add rule MULTIPLE_API_VERSION to validate if the default tag in readme.md contains multiple API version.